+++ packages/addons/tsconfig.test.json
@@ -1,11 +1,8 @@
 {
   "extends": "./tsconfig",
   "compilerOptions": {
+    "module": "ESNext",
     "sourceMap": true
   },
-  "include": [
-    "__tests__/**/*.ts",
-   // "src/**/*.ts"
-  ],
-  //"exclude": ["src"]
+  "include": [ "__tests__/**/*.ts" ]
 }
